+++ model/schema_get_test.go
@@ -1,0 +1,10 @@
+package model
+
+import "testing"
+
+func TestGet(t *testing.T) {
+	// TODO: Schema.Get
+	// TODO: Schema.GetAt
+	// TODO: Schema.GetOnly
+	// TODO: Schema.GetOnlyAt
+}